ai-mobile-stock-screener-week1/requirements.txt
psycopg2-binary==2.9.9
pandas==2.1.3
numpy==1.26.2
requests==2.32.3
